move player classes to own module
