+++ lib/play.js
@@ -4,6 +4,7 @@ const { checkForGameOver } = require("./checkWin");
 const { pickOrder, setPlayOrder } = require("./playOrder");
 const { playAgain } = require("./playAgain");
 const { Store } = require("./store");
+const { CpuPlayer, HumanPlayer } = require("./player");
 
 const playerTurn = async (store) => {
   const thisPlayer = store.nextPlayer;
