+++ lib/play.test.js
@@ -1,7 +1,8 @@
 // const { userTurn, cpuTurn, makeTurn } = require("./play");
 const { playerTurn } = require("./play");
 const rl = require("./rl");
-const { Store, HumanPlayer, CpuPlayer } = require("./store");
+const { Store } = require("./store");
+const { HumanPlayer, CpuPlayer } = require("./player");
 
 jest.mock("./rl");
