+++ lib/player.test.js
@@ -1,4 +1,5 @@
-const { Player, HumanPlayer, CpuPlayer, Store } = require("./store");
+const { Store } = require("./store");
+const { Player, HumanPlayer, CpuPlayer } = require("./player");
 const rl = require("./rl");
 
 jest.mock("./rl");
