+++ lib/store.test.js
@@ -1,5 +1,5 @@
-const { describe, expect } = require("@jest/globals");
-const { Store, CpuPlayer, HumanPlayer } = require("./store");
+const { Store } = require("./store");
+const { CpuPlayer, HumanPlayer } = require("./player");
 
 describe("Store methods can set values", () => {
   let consoleOutput = [];
